#30 glue code for fe and be

+++ be/src/route/v1/DELETE-role-from-group.route.ts
@@ -2,39 +2,47 @@ import {NextFunction, Router, Request, Response} from "express";
 import {Registry} from "../../registry";
 import {validateJwtMiddlewareFn, validateMiddlewareFn} from "./common-middleware";
 import {check} from "express-validator";
-import {doInDbConnection, QueryResponse} from "../../db";
+import {doInDbConnection, QueryA, QueryResponse} from "../../db";
 import {PoolConnection} from "mariadb";
-import {DeleteRoleFromGroupResponse} from "../../model/role.model";
+import {makeApiError, makeApiErrorObj} from "../../util";
+import {ApiResponse} from "../../model/response.model";
 
 const httpAction: any[] = [
     [
         check('groupId').exists().isNumeric(),
-        check('roleId').exists().isNumeric()
+        check('roleName').exists()
     ],
     validateMiddlewareFn,
     validateJwtMiddlewareFn,
     async (req: Request, res: Response, next: NextFunction) => {
 
         const groupId: number = Number(req.params.groupId);
-        const roleId: number = Number(req.params.roleId);
+        const roleName: string = req.params.roleName;
 
         await doInDbConnection(async (conn: PoolConnection) => {
+            const qs: QueryA = await conn.query(`SELECT ID FROM TBL_ROLE WHERE NAME = ?`, [roleName]);
+            if (qs.length <= 0) {
+                res.status(400).json(makeApiErrorObj(
+                    makeApiError(`Invalid role ${roleName}, roleName not found in system`, 'roleName', roleName, 'System')
+                ));
+                return;
+            }
+            const roleId: number = qs[0].ID;
             const q: QueryResponse = await conn.query(`
                 DELETE FROM TBL_LOOKUP_GROUP_ROLE WHERE GROUP_ID = ? AND ROLE_ID = ?
             `, [groupId, roleId])
 
             res.status(200).json({
                 status: (q.affectedRows ? 'SUCCESS' : 'ERROR'),
-                message: (q.affectedRows ? `Role ${roleId} deleted` : `Role ${roleId} is not in group ${groupId}`),
-                entriesDeleted: q.affectedRows
-            } as DeleteRoleFromGroupResponse)
+                message: (q.affectedRows ? `Role ${roleName} deleted from group ${groupId}` : `Role ${roleName} is not in group ${groupId}`),
+            } as ApiResponse)
         });
     }
 ];
 
 
 const reg = (router: Router, registry: Registry) => {
-    const p = `/group/:groupId/role/:roleId`;
+    const p = `/group/:groupId/role/:roleName`;
     registry.addItem('DELETE', p);
     router.delete(p, ...httpAction);
 }
